backend/requirements.txt
fastapi[standard]
alembic
asyncpg
sqlalchemy
pydantic-settings
python-jose
bcrypt==4.0.1
passlib[bcrypt]
